added javadocs.
